+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.24.0 (2023-03-06)
+
+- Add support for custom first question
+- Fix EditText background for Flutter
+- Fix AlertDialog theme for Flutter
+- Add checks for driver picklist fields
+- Remove unused code
+
 ## 2.23.0 (2023-02-09)
 
 - Add support for Driver Picklist feature